--- components/asset-controller/pkg/apis/objectstore/group.go
@@ -1,2 +1,0 @@
-// Package objectstore contains objectstore API versions
-package objectstore
