--- components/asset-controller/pkg/apis/objectstore/v1alpha1/doc.go
@@ -1,7 +1,0 @@
-// Package v1alpha1 contains API Schema definitions for the objectstore v1alpha1 API group
-// +k8s:openapi-gen=true
-// +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=github.com/kyma-project/kyma/components/bucket-controller/pkg/apis/objectstore
-// +k8s:defaulter-gen=TypeMeta
-// +groupName=objectstore.kyma-project.io
-package v1alpha1
