+++ components/assetstore-controller-manager/.gitignore
