+++ components/assetstore-controller-manager/Dockerfile
