+++ components/assetstore-controller-manager/Gopkg.toml
