+++ components/assetstore-controller-manager/config/crds/assetstore_v1alpha1_bucket.yaml
@@ -4,9 +4,9 @@ metadata:
   creationTimestamp: null
   labels:
     controller-tools.k8s.io: "1.0"
-  name: buckets.objectstore.kyma-project.io
+  name: buckets.assetstore.kyma-project.io
 spec:
-  group: objectstore.kyma-project.io
+  group: assetstore.kyma-project.io
   names:
     kind: Bucket
     plural: buckets
@@ -30,6 +30,7 @@ spec:
           properties:
             policy:
               enum:
+              - ""
               - none
               - readonly
               - readwrite
