+++ components/assetstore-controller-manager/config/default/kustomization.yaml
