+++ components/assetstore-controller-manager/config/default/manager_image_patch.yaml
