+++ components/assetstore-controller-manager/config/default/manager_prometheus_metrics_patch.yaml
