+++ components/assetstore-controller-manager/config/manager/manager.yaml
