+++ components/assetstore-controller-manager/config/rbac/auth_proxy_role_binding.yaml
