+++ components/assetstore-controller-manager/config/rbac/auth_proxy_service.yaml
