+++ components/assetstore-controller-manager/config/rbac/rbac_role.yaml
@@ -5,27 +5,7 @@ metadata:
   name: manager-role
 rules:
 - apiGroups:
-  - apps
-  resources:
-  - deployments
-  verbs:
-  - get
-  - list
-  - watch
-  - create
-  - update
-  - patch
-  - delete
-- apiGroups:
-  - apps
-  resources:
-  - deployments/status
-  verbs:
-  - get
-  - update
-  - patch
-- apiGroups:
-  - objectstore.kyma-project.io
+  - assetstore.kyma-project.io
   resources:
   - buckets
   verbs:
@@ -37,7 +17,7 @@ rules:
   - patch
   - delete
 - apiGroups:
-  - objectstore.kyma-project.io
+  - assetstore.kyma-project.io
   resources:
   - buckets/status
   verbs:
