+++ components/assetstore-controller-manager/config/rbac/rbac_role_binding.yaml
