+++ components/assetstore-controller-manager/config/samples/assetstore_v1alpha1_bucket.yaml
