+++ components/assetstore-controller-manager/pkg/apis/assetstore/group.go
@@ -1,0 +1,2 @@
+// Package assetstore contains assetstore API versions
+package assetstore
