+++ components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha1/bucket_types_test.go
