+++ components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha1/doc.go
@@ -1,0 +1,7 @@
+// Package v1alpha1 contains API Schema definitions for the assetstore v1alpha1 API group
+// +k8s:openapi-gen=true
+// +k8s:deepcopy-gen=package,register
+// +k8s:conversion-gen=github.com/kyma-project/kyma/components/bucket-controller/pkg/apis/assetstore
+// +k8s:defaulter-gen=TypeMeta
+// +groupName=assetstore.kyma-project.io
+package v1alpha1
