+++ components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha1/v1alpha1_suite_test.go
