+++ components/assetstore-controller-manager/pkg/apis/assetstore/v1alpha1/zz_generated.deepcopy.go
