+++ components/assetstore-controller-manager/pkg/buckethandler/automock/minio_client.go
