+++ components/assetstore-controller-manager/pkg/controller/controller.go
